rendering everything but light on wrong side

+++ srcs/shaders/hit_shader.c
@@ -22,12 +22,11 @@ t_vec4	hit_shader(t_scene *scene, t_payload *payload)
 
 	if (payload->light_dist > 0)
 	{
-		// payload->hitnorm = vec3_multf(payload->hitnorm, -1);
 		DEBUG_ONLY(print_vec3(payload->hitnorm, "hitnorm"));
 		
 		intensity = vec3_dot(payload->light_dir, payload->hitnorm);
 		DEBUG_ONLY(printf("intensity: %f\n", intensity));
-		intensity = fminf(fmaxf(intensity, 0.0), 1.0);
+		intensity = fmin(fmax(intensity, 0.0), 1.0);
 		DEBUG_ONLY(printf("intensity: %f\n", intensity));
 		
 		col = vec4_multf(scene->light.colour, intensity);
